win8.1,installer: Force Windows activation
I've noticed that when installing Windows 8.1 the activation is not done
by the end of the unattended installation and, actually, the serial
number shown in the Activate Windows program is different than the one
used to perform the installation. Also, when tring to activate the
software it fails with the "Windows can't be activate. Try again later"
message and no useful details about the error.

So, in order to avoid this issue, let's force the activation in the
FirstLogon phase of the installation script by directly editing the
Slmgr.vbs options for Volume Activation.

For more info about the Slmgr.vbs Options for Volume Activation, please,
take a look on:
https://technet.microsoft.com/en-us/library/dn502540.aspx

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Zeeshan Ali (Khattak) <zeeshanak gnome org>