installer: Use l10n-language as UI language
Since commit 57b459e09 we set 'EN-US' as UI language instead of 'EN-GB'.
However, this hack is only necessary for Windows 7 and can cause issues
when installing newer versions of the Windows, as the installer,
somehow, gets confused by having different values set to UI language,
System Locale and User Locale and ends up prompting a window for letting
the user to set these values.

By limiting the hack to Windows 7, this patch fixes the installation of
'EN-GB' ISOs of Windows 8 and 8.1 (the ones tested for now).

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Zeeshan Ali (Khattak) <zeeshanak gnome org>