debian,installers: Enable ftp mirror
Currently, after completing the installation, the users are not able to
install any software in the VM as the only entry in
/etc/apt/sources.list is the installation media itself.

Also, desktop installation's profile is broken as qemu-guest-agent is
not part of the installation media and we have no mirror set up for
downloading the package from.

Let's fix these issues enabling the ftp mirror and removing the DVD from
the /etc/apt/sources.list.

Resolves:
https://bugzilla.redhat.com/show_bug.cgi?id=1336110
https://bugzilla.redhat.com/show_bug.cgi?id=1336111

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Zeeshan Ali (Khattak) <zeeshanak@gnome.org>