Revert "debian: Allow empty passwords"
The workaround that has been used so far doesn't work.

For the user account, the password is indeed removed in the end of the
installation, but then login in from GDM is impossible. Although it
works from a VT (depending on the installed version), this is not
something obvious that every user would try. So, requiring the user
password seems the best to do for now and when a proper solution is
found, the user password can be set to optional again.

For the root account, the password is not removed in the of the
installation and ends up being set as "dummyPa55w0rd", something that
the user would never guess, unless they have access to the libosinfo
code. So, requiring the admin password seems the best to do for now and
when a proper solution is found the admin password can be set to
optional again.

This reverts commit 087a8f8a23895e2d792b0f6d26a9d0c01a6c7e87.

 Conflicts:
	data/install-script/debian.org/debian-preseed-desktop.xml.in

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>

Resolves:
https://bugzilla.redhat.com/show_bug.cgi?id=1336108
https://bugzilla.redhat.com/show_bug.cgi?id=1336109

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Zeeshan Ali (Khattak) <zeeshanak@gnome.org>