Disable -Wunsafe-loop-optimizations warnings
This warning option is not reporting broken code. Rather it is merely
identifying loops which GCC would not be able to optimize fully when
using -funsafe-loop-optimizations.  In GCC 7 this triggers on the
loader code

osinfo_loader.c: In function 'osinfo_loader_process_list':
osinfo_loader.c:2127:11: error: missed loop optimization, the loop counter may overflow [-Werror=unsafe-loop-optimizations]
     while (tmp && *tmp) {
           ^

The warning just needs to be turned off since the code is not broken,
and we don't care about the optimization level

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>