Add 'eject-after-install' parameter to OsinfoMedia
If the media is an installer, thus specifies whether the media should be
ejected after the installation process. Default value is true.

This is mainly needed for applications like GNOME Boxes (and maybe
virt-install) to be able to decide whether the media should be ejected
or not when the final reboot happens during its installation process.

The latter case may happen when the installer leaves some packages to be
installed after rebooting the OS for the last time.

The issue solved by this patch is a corner-case faced when adding the
install scripts for SLES, as during its installation only one reboot is
performed (so, installer-reboots attribute doesn't help us) and the media
must not be ejected after the reboot in order to finish the installation.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>
Acked-by: Daniel P. Berrange <berrange@redhat.com>