Switch from libsoup to libcurl
libsoup is used to check the validity of URLs in distributions in
osinfo-db; OTOH it supports only HTTP(S), so this limits the checks to
that protocol.

To overcome this limitation, switch to libcurl: while it requires
slightly more code to do the same task, it provides a bit more
flexibility, and support for other protocols.  No version check is
performed, since the APIs used are old enough.

Adapt also the README, and the packaging files.

Acked-by: Christophe Fergeau <cfergeau@redhat.com>