build: make curl optional, and only for tests
Look for curl only when the tests are enabled, and as optional build
dependency.  If curl is missing, then the two network tests are not
built.  Also, limit the use of curl CFLAGS/LIBS only for the tests that
really need it.

As side result, it is possible to drop the curl BuildRequires from the
mingw spec, as it builds with --enable-tests=no.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>