test-isodetect: Iterate over all isodata/ subdirs
When adding support for a new vendor, test-isodetect.c needs to be
updated to list it, which is too easy to forget.
This commit iterates over subdirs in tests/isodata/ and uses these as
the vendors to test for. This means no further modification of
test-isodetect.c is needed when adding a new vendor.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>