autogen: Don't hardcode --enable-gtk-doc
The usual behavior is for configure to detect the availability
of gtk-doc, along with other optional features, and only enable
it if found.

Hardcoding it this way is causing issues for MinGW builds,
where pkg-config is looking in the MinGW-specific paths and
can't locate the native gtk-doc.pc file, thus concluding
gtk-doc must not be present and aborting the build altogether
instead of simply skipping API reference generation.

Both 'make distcheck' (through the DISTCHECK_CONFIGURE_FLAGS
variable) and the prepare-release.sh script are already
enforcing the availability of gtk-doc when preparing releases,
so it's safe to use the default behavior for regular builds.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>