Introduce OsinfoImage object
OsinfoImage object has been created to represent pre-installed images
distributed by the OSes.

We've decided to go for the simplespet approach possible that's just
exposing the Image URL for an existing OS and then Apps that want to
deal with the Image itself can just use libguestfs for that as they
already do a quite good job on this field.

https://gitlab.com/libosinfo/osinfo-db/issues/10

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Pino Toscano <ptoscano@redhat.com>