os: Support inherited resources
As resources are pretty much the same for new releases of a same OS,
having them explicitly set is a source of copy-and-paste error and
redundant information in our db.

The approach taken to properly support inheritance of OSes is a
explicitly set attribute to the resources indicating that they should be
inherited.

Here's an example of how the XML will look like:
foo:
  <resources arch="all">
    <minimum>
      <n_cpus>1</n_cpus>
      <storage>1</storage>
    </minimum>
  </resources>

bar (which derives-from/clones foo):
  <resources arch="all" inherit="true">
    <minimum>
      <n_cpus>3</n_cpus>
      <ram>3</ram>
    </minimum>
  </resources>

When querying for bar's minimum resources, we'll get:
n_cpus: 3
cpu: -1 (not set, by any of the OSes)
ram: 3
storage: 1

https://gitlab.com/libosinfo/osinfo-db/issues/15

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>