image: Add "cloud-init" attribute to the image
Some images may support cloud init customization and for those we'd like
to explicitly indicate it.

The "cloud-init" attribute is similar to the "live" attribute for
medias, has its default value as FALSE and can be optinionally set in
the XML.

This commit also extends the basic image tests to also cover the
cloud-images attribute.

https://gitlab.com/libosinfo/osinfo-db/issues/10

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Pino Toscano <ptoscano@redhat.com>