tests: Fix `make check` breakage for FreeBSD 11
Let's take advantage of the newly added libosinfo-impl.la that contains
all the symbols that we need and link our tests against it.

By doing so, we can remove the referrence to osinfo_product.c when
building test-os, which has been causing test-os binary to contain 2
copies of the very same code, one linked in the test and the other from
libosinfo.so.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Christophe Fergeau <cfergeau@redhat.com>
Tested-by: Andrea Bolognani <abologna@redhat.com>