install-script: Add _(get|set)_preferred_injection_method()
Those new methods are going to be used to tell the install-scripts
whether the injection-method that's going to be used is. We have to do
so as the command-line may be different depending on the
injection-method used, for example:
- fedora using cdrom, disk or floppy: ks=hd:/(vda|sda)/fedora.ks
- fedora using initrd: ks=file:/fedora.ks

It's important to mention that although the methods are taking GFlags,
those are treated as GEnum and only one value is expected to be set, as
mentioned in the documentation.

Also, mind that the usage of osinfo_entity_set_param() to store the
nick of the OSINFO_INSTALL_SCRIPT_INJECTION_METHOD_* is intentional as
the nick is exactly what's going to be used in the install-scripts to
generate the proper command-line.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>