tree: cleanup non-fatal errors in load_key_info()
There are errors which are not fatal and just ignored in load_keyinfo.
However, as those have not been cleaned up, we could see messages like:
(lt-osinfo-detect:20658): GLib-WARNING **: GError set over the top of a
previous GError or uninitialized memory.
This indicates a bug in someone's code. You must ensure an error is NULL
before it's set.
The overwriting error message was: Key file does not have key “boot.iso”
in group “images-x86_64”

In order to avoid this, let's just call g_clear_error() after situations
where an error may have been set but it can just be ignored.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Christophe Fergeau <cfergeau@redhat.com>