os: Deal with network-install resources
network-install resources represent the resources to be used during a
network installation. It may differs from the minimum resources as the
network installers (not only the medias, but the trees) may download a
huge amount of data to the RAM* consuming way more than the minimum or
recommended RAM for the OS.

*: anaconda downloads the whole stage2 of the installer into the RAM in
order to perform a network installation.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>