tests: add one test case per OS tested
The image, media & tree URIs test suites just register one test case for
all the work they do. As a result when they run they just sit there
printing no progress information for a long time.

This changes then to register one test case per OS in the database. As a
result we get progress information. An added benefit is that it is
possible to now run tests for individual OS by passing the -p argument.

  eg  test-mediauris -p /mediauris/uris/rhel3.9

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>