tests: Avoid re-using the curl handler on URIs tests
On CentOS, we've hit an issue that when a curl handler is re-used we end
up receiving malfunctioning responses from NetBSD CDN.

Dan investigated the issue found out that the problem is that libcurl in
CentOS 7 at least is not ignoring the body. Sometimes it closes the
keepalive connection so this doesn't matter, but other times it gives
us back the body that it should be ignoring.

In order to work the issue around, let's no re-use the curl handler but
always open a new one for each of the URIs.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>