tests: Reduce code duplication in URI tests
Let's introduce a common file that does the check of the URI for all the
URI related tests.

By doing this we can avoid code duplication while also avoiding to have
all the tests merged together into a single one, which is not so good as
limits the developer to run and wait for the whole set of tests instead
of having the ability to just run the a specific one.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>