media: Don't assume identifiers are null terminated
Identifiers as volume-id, application, publisher, and system are not
null terminated and cannot be assumed as so.

By assuming those are null terminated strings, libosinfo ends up not
counting the last character of a MAX_* string and, consequently, not
properly identifying medias that have their identifiers with the MAX_*
size.

One example is the ubuntu-18.04.1.0-live-server-amd64.iso media, which
has as volume-id 'Ubuntu-Server 18.04.1+ LTS amd64'. As the volume-id
has exactly 32 characters it's never been matched as when reading the
media's volume-id it'd be read as 'Ubuntu-Server 18.04.1+ LTS amd6'.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Christophe Fergeau <cfergeau@redhat.com>