spec: Adapt to use meson
The changes needed in the spec files are not so intrusive.

One thing that has to be mentioned is the requirement for
mingw-filesystem >= 107 (which is only present on Fedora 30+) in order
to generate the mingw rpms for osinfo-db-tools. This huge bump in the
version happened because the `%mingw_ninja_install` was only recently
introduced to mingw-filesystem package.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>