os: Move the "Since: " annotation
The "Since: " annotation should be *always* added to the end of the
comment block. More than that, multiplce "Since: " annotations cannot be
part of a same comment block and in order to solve that, let's avoid
adding a "Since: " in the end of the block, but explicitly mention when
each of the values were added.

By taking this approach, the following warnings can be avoided:
```
../osinfo/osinfo_os.h:90: Warning: Libosinfo: "@OSINFO_RELEASE_STATUS_RELEASED" parameter unexpected at this location:
 * @OSINFO_RELEASE_STATUS_RELEASED: A released OS.
    ^
../osinfo/osinfo_os.h:96: Error: Libosinfo: multiple "Since:" tags for identifier "OsinfoReleaseStatus":
 * Since: 1.4.0
   ^
../osinfo/osinfo_os.h:97: Warning: Libosinfo: "@OSINFO_RELEASE_STATUS_ROLLING" parameter unexpected at this location:
 * @OSINFO_RELEASE_STATUS_ROLLING: A rolling version of an OS and thus with no
    ^
```

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>