Silence -Wstring-plus-int
The warning is triggered because we're ignoring the "treeinfo-" part of
the OSINFO_TREE_PROP_TREEINFO_* string.

A possible way to fix that, instead of just disabling the warning, would
be doing something like:
```
const char *prop = OSINFO_TREE_PROP_TREEINFO_FAMILY;
prop += strlen("treeinfo-");
```

However, I sincerely do not think it's worth the verbosity of the
solution.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>