loader: Avoid conflicting object ids for extension files
Currently, when building the id of some of our objects, we do:
- Loop over the several objects of that type;
  - Concatenate the id of the parent object with the loop index;

By doing this, conflicts can be seen when using more than one extension
file, as osinfo-db does for Windows, where pre and post installable
drivers extension files are used and end up being loaded as we'd do:

- Loop over pre installable drivers;
  - Concatenate the id of the parent object with the loop index;

- Loop over post installable drivers;
  - Concatenate the id of the parent object with the loop index;

In order to avoid such issues, let's also use the filename as part of
the object id, avoiding then conflicts.

It's important to note that the conflict is cause by osinfo_list_add(),
which will replace the object in the list in case some other object with
the very same id is already stored there.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>