Add API to track distro cloud-init usernames
We already have cloud-init='true' so add a cloud-image-username='foo' to
go with it.

The property has been added to `OsinfoOs` instead of `OsinfoImage`
because it's not expected to change in a per image basis. Instead we can
define it once and inherit it, in the same way that it's done for
kernel-url-argument.

Fixes #39

Signed-off-by: Rafael Fonseca <r4f4rfs@gmail.com>