media/tree: Force "wget" user-agent in SoupSession
Spoiler alert, it'll sound insane ... because it is insane. Some servers
treat libsoup user-agent differently than how they treat wget
user-agent.

With the default agent set, Soup would end up downloading the file
containing the redirection information, instead of actually redirecting
to one of the addresses present in that file. While when forcing "wget"
as user-agent, the redirection happens as expected.

This issue was faced at least with CentOS[0] and Manjaro[1] links.

After a talk with Patrick Griffis, he mentioned that:
"I imagine there isn't too much of a problem pretending to be wget. I would
guess the assumption by these services is 'A CLI tool is downloading us and
not a browser' rather than 'wget has unique behavior'.".

Knowing that, let's force "wget" as user-agent on our side.

[0]: http://isoredirect.centos.org/centos/8/isos/x86_64/CentOS-8.1.1911-x86_64-dvd1.iso
[1]: https://osdn.net/dl/manjaro/manjaro-gnome-19.0.2-200312-linux54.iso

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>