db: Match "rolling" distros only as a fallback
Let's only match "rolling" distros as a fallback. The reason we need to
do so is to properly deal with distros as Manjaro, which has a mix of
versioned and rolling entries.

By matching the "rolling" only as a fallback, we ensure we'll have
exhausted all the "versioned" possibilites in the first place.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>