prepare-release: Drop RESULTS
This is a leftover from when the script was used to perform
automated tests: in that context, singling out the output of the
test suite made sense, but now we're no longer particularly
interested in it.

As a bonus, this allows us to drop a non-trivial chunk of fairly
obscure POSIX shell compatibility code.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>