prepare-release: Use DESTDIR when installing
DESTDIR and --prefix are similar but serve distinct purposes: in
particular, the latter is intended for modifications that will
persist after installation, eg. storing binaries under /usr/bin
instead of /usr/local/bin, while the former is just a temporary
change that can be easily reverted, eg. stashing files in an
intermediate, unprivileged directory while building a package.

Our needs during the release process are closer to what DESTDIR is
designed for, so let's use that.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>