ci: Split off mingw-environment
Follow the environment/script split, like we do everywhere else.

We can have a single environment that works both for mingw32 and
mingw64 builds because the only difference between the two is
which ABI they target, and we can obtain that piece of information
from the $ABI environment variable that's baked into the container
images.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>