ci: Introduce multilib
This is a bit yucky, but it's only temporary: like the target ABI,
this information is tied to the container image and as such should
be baked into it. When that happens, we can simply drop this
altogether, but in the meantime this bit of scripting allow us to
have a single environment instead of two.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>