build-aux: use gmake for syntax-check rules
Previously on FreeBSD the syntax-check rules would be a no-op because
its non-GNU make wouldn't see the GNUmakefile and would report no work
to do:

$ MESON_SOURCE_ROOT=`pwd` build-aux/syntax-check
`syntax-check' is up to date.

Since we moved everything into a syntax-check.mk that we call explicitly
though, we're the non-GNU make tries to execute rules that depend on
GNU make features.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>