ci: define a CI manifest file
This is to be used with the command "lcitool manifest ci/manifest.yml"
to re-generate all existing dockerfiles and gitlab CI config.

In a slight change to the existing config, the manifest will always
use the pre-build osinfo-db-tools package. Nothing in the libosinfo
CI tasks, has a need to exercise latest git master for osinfo-db-tools.
The latter is covered fine by osinfo-db CI jobs.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>