Skip over entity reference checks if loading failed
If we fail to parse one of the XML files, we abort loading the
database. We then carry on to do an entity reference check which
will almost certainly fail due to not having loaded many XML
files. We need to skip over the entity reference check on failure
to avoid spamming the user with bogus errors.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>