spec: require gobject-introspection
Latest pygobject3 in Fedora dropped dependency on gobject-introspection
as they are now using implementation from GLib2. This causes error when
user tries to import libosinfo in python:

    import gi
    from gi.repository import Libosinfo

fails with error:

    ImportError: Typelib file for namespace 'libxml2', version '2.0' not found

The libxml2-2.0.typelib file is installed by gobject-introspection so we
need to depend on it to fix this import error.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2412451
Signed-off-by: Pavel Hrdina <phrdina@redhat.com>